+++ src/main/java/com/unstoppabledomains/exceptions/dns/DnsExceptionCode.java
@@ -1,0 +1,6 @@
+package com.unstoppabledomains.exceptions.dns;
+
+public enum DnsExceptionCode {
+  DnsRecordCorrupted,
+  InconsistentTtl;
+}