+++ src/main/java/com/unstoppabledomains/resolution/contracts/cns/ProxyReader.java
@@ -1,6 +1,6 @@
 package com.unstoppabledomains.resolution.contracts.cns;
 
-import com.unstoppabledomains.exceptions.NamingServiceException;
+import com.unstoppabledomains.exceptions.ns.NamingServiceException;
 import com.unstoppabledomains.resolution.contracts.BaseContract;
 
 import java.math.BigInteger;