+++ src/main/java/com/unstoppabledomains/resolution/dns/DnsRecord.java
@@ -1,0 +1,16 @@
+package com.unstoppabledomains.resolution.dns;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.ToString;
+import lombok.EqualsAndHashCode;
+
+@Getter
+@AllArgsConstructor
+@ToString
+@EqualsAndHashCode
+public class DnsRecord {
+  private DnsRecordsType type;
+  private int ttl;
+  private String data;
+}